cii: 2018: reports: Report of Linux Kernel bugs resolved in March
Add detailed report of Linux kernel bugs resolved in March 2018.

Most of the bugs marked as NA were found with the help of Coccinelle
and they might have a Coverity ID in the data base.

Also, some of the patches resolve more than one Coverity issue of the
same or different type.

Signed-off-by: Gustavo A. R. Silva <gustavo@embeddedor.com>