cii: reports: annual: Annual report of bugs and issues resolved 2017-2018
Add detailed report of all bugs and issues resolved from May 2017 to
end of April 2018.

Most of the bugs marked as NA were found with the help of Coccinelle
and they might have a Coverity ID in the data base.

Also, some of the patches resolve more than one Coverity issue of the
same or different type.

This report will be keep up-to-date until the end of April 2018.

Signed-off-by: Gustavo A. R. Silva <gustavo@embeddedor.com>